+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 - `file/reader`: Make `read()` returns the exact data
   - Currently, it always returns `[]` due to an implementation mistake.
+- `file/writer`: Flush buffer after the data written ended.
 
 ### Unreleased: 0.3.0 – Tests